+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2019-03-20
+### Added
+- Input for cluster scaling variables
+- Input for public subnets (cluster tagging)
+- Input for installation of Helm
+### Changed
+- Incremented version of EKS module
+
 ## [0.3.0] - 2019-03-06
 ### Changed
 - Common tag set created